# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "战鸽数据",
    "description": "Zhange Stats · Steam 游玩统计与圈子成员管理",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/exilium/status": {
      "get": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Status",
        "operationId": "exilium_status_api_exilium_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_roles",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Roles"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exilium/logs": {
      "get": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Logs",
        "description": "兼容占位：历史列表已弃用；今日状态见 status（读 *_checkin_logs 缓存）。",
        "operationId": "exilium_logs_api_exilium_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exilium/bind/password": {
      "post": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Bind Password",
        "operationId": "exilium_bind_password_api_exilium_bind_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExiliumBindPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exilium/bind/sms/send": {
      "post": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Bind Sms Send",
        "operationId": "exilium_bind_sms_send_api_exilium_bind_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExiliumBindSmsSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumBindSmsSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exilium/bind/sms": {
      "post": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Bind Sms",
        "operationId": "exilium_bind_sms_api_exilium_bind_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExiliumBindSmsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exilium/bind": {
      "delete": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Unbind",
        "operationId": "exilium_unbind_api_exilium_bind_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumStatusOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Update Bind",
        "operationId": "exilium_update_bind_api_exilium_bind_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExiliumBindUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exilium/checkin": {
      "post": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Checkin Now",
        "operationId": "exilium_checkin_now_api_exilium_checkin_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exilium/exchange": {
      "get": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Exchange Shop",
        "operationId": "exilium_exchange_shop_api_exilium_exchange_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumExchangeShopOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Do Exchange",
        "operationId": "exilium_do_exchange_api_exilium_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExiliumExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumExchangeResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exilium/score-logs": {
      "get": {
        "tags": [
          "exilium"
        ],
        "summary": "Exilium Score Logs",
        "description": "积分变动记录（对应官方 points_Log）。",
        "operationId": "exilium_score_logs_api_exilium_score_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExiliumScoreLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "role_uid": {
            "type": "string",
            "title": "Role Uid"
          },
          "role_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Name"
          },
          "channel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label",
            "default": ""
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "awards_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards Text"
          },
          "checkin_date": {
            "type": "string",
            "title": "Checkin Date"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_code",
          "game_name",
          "role_uid",
          "status",
          "checkin_date",
          "checked_at"
        ],
        "title": "CheckinLogOut",
        "description": "签到记录公共结构（森空岛 / 塔吉多等共用）。"
      },
...
      "CheckinResponse": {
        "properties": {
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          },
          "ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "CheckinResponse"
      },
      "CheckinResultItem": {
        "properties": {
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "role_uid": {
            "type": "string",
            "title": "Role Uid"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "channel_name": {
            "type": "string",
            "title": "Channel Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "awards_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards Text"
          },
          "extra_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Text"
          }
        },
        "type": "object",
        "required": [
          "game_code",
          "game_name",
          "role_uid",
          "role_name",
          "channel_name",
          "status",
          "message"
        ],
        "title": "CheckinResultItem"
      },
...
      "ExiliumBindPasswordRequest": {
        "properties": {
          "account": {
            "type": "string",
            "maxLength": 128,
            "minLength": 5,
            "title": "Account"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 4,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "account",
          "password"
        ],
        "title": "ExiliumBindPasswordRequest"
      },
      "ExiliumBindSmsRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "captcha": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Captcha"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "captcha"
        ],
        "title": "ExiliumBindSmsRequest"
      },
      "ExiliumBindSmsSendRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "graph_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Code"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "ExiliumBindSmsSendRequest"
      },
      "ExiliumBindSmsSendResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "验证码已发送"
          },
          "need_graph_captcha": {
            "type": "boolean",
            "title": "Need Graph Captcha",
            "default": false
          },
          "graph_captcha_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Captcha Image"
          }
        },
        "type": "object",
        "title": "ExiliumBindSmsSendResponse"
      },
      "ExiliumBindUpdate": {
        "properties": {
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 23.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 59.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          }
        },
        "type": "object",
        "title": "ExiliumBindUpdate"
      },
...
      "ExiliumExchangeRequest": {
        "properties": {
          "exchange_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Exchange Id"
          }
        },
        "type": "object",
        "required": [
          "exchange_id"
        ],
        "title": "ExiliumExchangeRequest"
      },
      "ExiliumExchangeResultOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExiliumExchangeItemOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ExiliumExchangeResultOut"
      },
      "ExiliumExchangeShopOut": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score",
            "default": 0
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExiliumExchangeItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "ExiliumExchangeShopOut"
      },
      "ExiliumRoleOut": {
        "properties": {
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "channel_name": {
            "type": "string",
            "title": "Channel Name"
          }
        },
        "type": "object",
        "required": [
          "game_code",
          "game_name",
          "uid",
          "role_name",
          "channel_name"
        ],
        "title": "ExiliumRoleOut"
      },
...
      "ExiliumScoreLogOut": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/ExiliumScoreLogItemOut"
            },
            "type": "array",
            "title": "List"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "default": 50
          }
        },
        "type": "object",
        "title": "ExiliumScoreLogOut"
      },
      "ExiliumStatusOut": {
        "properties": {
          "bound": {
            "type": "boolean",
            "title": "Bound"
          },
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          },
          "phone_mask": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Mask"
          },
          "bound_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bound At"
          },
          "last_checkin_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin At"
          },
          "last_checkin_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Date"
          },
          "last_checkin_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Ok"
          },
          "last_checkin_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Summary"
          },
          "token_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Ok"
          },
          "token_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Error"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/ExiliumRoleOut"
            },
            "type": "array",
            "title": "Roles"
          },
          "today_results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Today Results"
          },
          "today_logs": {
            "items": {
              "$ref": "#/components/schemas/CheckinLogOut"
            },
            "type": "array",
            "title": "Today Logs"
          }
        },
        "type": "object",
        "required": [
          "bound"
        ],
        "title": "ExiliumStatusOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}